@@ -15,7 +15,7 @@
         }
     ],
     "require": {
-        "php": "^7.3 || ^8.0",
+        "php": "^7.4 || ^8.0",
         "laminas/laminas-feed": "^2.14",
         "laminas/laminas-http": "^2.14",
         "composer/installers": "^1.0"
