@@ -12,14 +12,14 @@ class RemoveUniqueConstraintFromItems extends Migration
 {
     public function up(): void
     {
-        Schema::table('vdlp_rssfetcher_items', static function (Blueprint $table) {
+        Schema::table('vdlp_rssfetcher_items', static function (Blueprint $table): void {
             $table->dropUnique('item_id_unique');
         });
     }
 
     public function down(): void
     {
-        Schema::table('vdlp_rssfetcher_items', static function (Blueprint $table) {
+        Schema::table('vdlp_rssfetcher_items', static function (Blueprint $table): void {
             $table->unique('item_id', 'item_id_unique');
         });
     }
